Added restriction to Create, Read, Update and Delete resource
Changed the function name from get(), getAll(), create(), update() and delete() to getForUser(), getAllForUser(), createForUser(), updateForUser() and deleteForUser()

Added the authenticated user_id as a parameter and argument to all the renamed functions.

Included the authenticated user_id to the query string to restrict the query to retrieve only the resource for the authenticated user.