Cached the Database connection
To prevent multiple connection request to the Database.

In the Database class the Database connection was cached to the $_conn private property.

The $_conn property was set to a NULL value by default.

The PDO type was prefixed with a ? since type declaration is being used.

Then check if the Database connection stored in the $_conn property is null.