Created a login page
Created login.php

declared strict type=1

Restricted the request method too "POST".

For all other request methods a 405 http_response_code and a header "ALLOW: post" will be sent as the response.

If the form is submitted without a username or password a 400 http_response_code and an array with the message "Missing login credentials" will first be converted to JSON using the json_encode" before being sent as the response.