Updated Login validation with access token
In the UsersGateWay Class the function getUser() was renamed to getUserByUserName().

Instead of using the id to retrieve the users data the username is used to retrieve the data.

Then the password used to login is compared to the hashed password
stored in the database.

If both username and password are valid Access Token is created storing the users data, ex. "id" or "name."

This Access Token is created from an associative array that is encrypted using json_encode() and  base64_encode().

The Access Token is include with the headers Authorization:Bearer.

When retrieving users data the Access Token is decoded using base64_decode() , validated, then decoded from JSON, validated before storing the users "id" into the users_id property and returned to be used to retrieve the users data.