Created a Class to configure JSON Web Token (JWT)
Created a Class, JWTCodec.

Created another baseUrlEncode() function to convert the JWT header value into URL safe values first using the base64_encode , converting the string to an encoded data with MIME base64. Then using the str_replace function to replace any "+", "/", "=" to
"-", "_", "", making it url safe.
Created an encode() function that excepts an array of users credentials for the JWT payload and the url safe value is returned.

In the encode() function the the header and the payload is converted using the JSON encoded function.

After the header and the payload is converted into JSON it is passed to the baseUrlEncode function.

Then we create an random,  generated 256 bit 32 Byte Hex to be used in the hash_hmac() function with the sha256 algorithm and the header.
then passed into the baseUrlEncode function and returned.