Update JWTCodec.php
A decode() method was added to the JWTCodec class.

The decode() method will reverse the process the encode() method.

First receiving, as a dependency, the Access Token generated when the user successfully logs in and the users information is encoded.

Then the passed in encoded token is decoupled using the preg_match() function to extract the "signature" compare if the signature matches the signature generated when the user first logged in.

If both the signatures, that was  passed in and the signature created when the user successfully logged in, matches the "payload" extracted from the decoupled access token using the preg_match() function converted back to a associative array and returned.