Updated the .env file, JWTCodec Class and Login file
A constructer method was created in the JWTCodec class to receive as a dependency the "256 encrypted hex key".

The "256 encrypted hex key" is stored in the ".env" file as SECRET_KEY.

In the "login.php" file the "256 encrypted hex key" is passed in as an argument where the "new JWTCodec($_ENV['SECRET_KEY'])" object is created.