Updated the Auth class
In the authenticateAccessToken() method the "base64__decode()" and "json_decode()" methods were disabled since we the "256 encrypted hex key" and not "Access Token" is being used to retrieve the logged in users "id".

To implement this a new "_JWTCodec" private property was created and used to store the JWTCodec class that is passed in as a the 2nd argument in the  "__construct(UsersGateway, JWTCodec)" method and assigned to the "_JWTCodec" private property.

In the authenticateAccessToken()  method first the "256 encrypted hex key" is retrieved  from the "HTTP_AUTHORIZATION" header.

Then the value stored in the "HTTP_AUTHORIZATION" header is validated and converted into an array using the "preg_match" function .

Then the "decode()" function is called from the "JWTCodec" class.

And the 2nd array value in the "HTTP_AUTHORIZATION" header, which contains the "256 encrypted hex key", that was created from the "preg_match()", is passed in as an argument to the "decode($key)" method.

Then the decoded associative array value returned from the "decode()" function is stored in the "$data" variable.

The "$data["sub"] value stored in the "$data" array is then stored in the "private "_user_id" property and then returned with users id.

Then a "try{}catch{}" is used to catch if the secret token is valid.

In the "index.php" file now we create a new "JWTCodec()" object with the "Secret Key" and pass it as the 2nd  argument into the "Auth" class.