Created a InvalidSignatureException class
In the "InvalidSignatureException"class we extend the "Exception" class.

We do not add any logic to this class because the extended "Exception" class contains all the logic needed.

In the "JWTCodec" class when check if the (JWT) signature matches, if it does not match instead of throwing the standard " new Exception" object we throw the "new InvalidSignatureException" object.

In the Auth class we add a "catch" block to the "try/catch" block and if the (JWT) signature doesn't match the "InvalidSignatureException" Object will be true and a 402  HTTP Response is sent with a JSON message of "The signatures do not match" and a value of "FALSE" is returned.