Added expiring and refreshing Access Tokens
In the "login.php" page another key/value pair was added to the "payload" associative array.

The key added is named "exp" and the value is a php "time()" function set to 5 minutes.

A "TokenExpiredException" class was created to create another custom "Exception" handler. Like the "InvalidSignatureException"class no logic is required.

In the "JWTCodec" class after decoding the (JWT) "payload" we run a "if" statement to check if the "exp" time() value has expired.

If the "exp" time() value has expired  a new "TokenExpiredException"
"Exception" is thrown and a HTTP status is set to 401 and a JSON encoded message of "your login has expired" is set.