Generate a refresh token when the Access Token expires
created a new php filed called "refresh.php".

This file is configured to retrieve the value of the "Refresh Token" from the request and if it is invalid return a HTTP error response if it is not present.

This file will perform the some of the same actions like the "login.php" file.
So we USE the "enabling strict type", "require" the "bootstrap.php" file, restrict the "Method" to access the "refresh.php" end point to "POST", retrieve the the "Refresh Token"  from the request body as an associative array. Check if the array key "token" exist.  If it doesn't exist we set a HTTP 401 response, return a JSON encoded message and exit the script.

If the if the array key "token" exist we insatiate a new "JWTCodec" object with the "Secret Key" passed in a s an argument.

We call the "decode" method from the "JWTCodec" object and pass the "Refresh Token" in as a argument.

we run a "try/catch" to check if the "Refresh Token" is valid. If it is not valid we set a HTTP 401 response, return a JSON encoded message and exit the script.

If the "Refresh Token" is valid we retrieve the users "id".