The refresh token is generated in the login.php page
iIn the "login.php" file a "$refresh_token" associative array is created storing 2 parameters.

The first key as "sub "and and the value as the users "id".

The second key as "exp" and the value as the  "time()" function setting the expiration time for 5 days.

Then encoding the "$refresh_token" using the "encode" method in the "JWTCodec" class.