Updated the UsersGateway class
A new method "getUserById" was added to the "UsersGateway" class .

In the "refresh.php" endpoint file we perform the same action as the "login.php" endpoint file.

After retrieving the decoded users "id" from the "$refresh_payload" variable we instantiate a new "Database" object pass it as an argument to the instantiated "UsersGateway" class.

Then call the "getUserByRefreshTokenId" method and pass the "$refresh_user_id" and store the return value in the "$user" variable.

Check if the returned value stored in the "$user" variable is false or has the users data.

If the value stored in the "$user" variable is false the HTTP response header is set to 401, a JSON response message  is set to "Invalid Authentication" and the script is exited