Issue a new Access and Refresh Token
A "Token.php" file was created in the API folder to extract out logic that is being used in the "login.php" file to create a new Access and Refresh Token.

This logic is put into the ""token.php file and the "refresh.php" file.

In both the "login.php" and the "refresh.php" file require the "token.php" file. to generate the new "Access Token" and the "Refresh Token".