Create functionality to store the "refresh token" and expiration time in the database
In the database a new table "refresh_token" was created to store a  "hashed" version of the "refresh_token" and the expiration time in the database when the user  logs into the API.

A new Class file "RefreshTokenGateway" was created in the source folder to create and store a new "refresh_token" and the expiration time in the database.

In the "tokens.php" file the refresh_token payload expiration time was stored in the variable "$refresh_token_expiry" to use the expiration value in the "$refresh_token" payload and also as a argument when calling the "$createRefreshToken" method in the "RefreshTokenGateway" class.

In the "login.php" file a new "RefreshTokenGateway" was created and the "Database" connection along with the "secret key", which is stored in the ".env" filed, is passed in as the arguments.

Then the "createRefreshToken()" method is called and the "$encoded_refresh_token" along with  the "$refresh_token_expiry" is passed  in as the arguments from the "tokens.php" file.