Delete and Update the "Refresh_Token"
In the "refresh.php" file we add the same functionality used in the "login.php" file to store the "$encoded_refresh_token" and the "$refresh_token_expiry" in the database.

But before we call the "createRefreshToken()" first we create a "getRefreshToken()" method in the "RefreshTokenGateway" class to retrieve the current "refresh_token" from the database.

in the "refresh.php" file we move and call the "RefreshTokenGateway" object right after calling the new  "Database" object.

Then we call the "getRefreshToken()" method and pass the encoded  "token" associative array key, decoded from JSON, from the "input" request body.

Then validate if the "refresh_token" exists in the database.

If the return value returns "false" we set the "HTTP" response to (400) and set the message to "Invalid token (not on whitelist)".

If the return value contains the "refresh_token" table values then we call the "deleteRefreshToken()" method to delete the existing refresh token fro the database.

And lastly we call the "createRefreshToken()" method passing in the "encoded_refresh_token" and the "refresh_token_expiry" as arguments to create a new "refresh_token" value and a new "expires_at" expiration time.