Created a "logout.php" file to remove "Refresh_Token" from database
In the "src" folder the "logout.php"  is created.

In the "logout.php" file we are using most of the same code base in the "refresh.php" file.

The lines of code we are not using are  "$refresh_user_id = (int) $refresh_payload["sub"]" to validate if the users "id" is valid.

Remove the require field that shares the logic to encode the new "Access Token Payload" and a "Refresh Token Payload".

Remove the "$createRefreshToken = $RefreshTokenGateway->createRefreshToken()" that is responsible for storing the encoded "refresh_token" and the "expiry" time in the database.

The "$delRefreshToken = $RefreshTokenGateway->deleteRefreshToken()" is left to delete the "refresh_token" from the database.