Created "delete-expired-refresh-tokens.php" to delete expired tokens & updates PHP version
In the "RefreshTokenGateway.php" file we add the "deleteExpiredToken" to delete the "refresh_token" from the database if the "expires_at" is less than the "UNIX_TIMESTAMP()" and if it is less than the "UNIX_TIMESTAMP()" it will return a "row_count()".

In the root folder "delete-expired-refresh-tokens.php" file was created to delete expired "refresh_tokens".

First we "delcare(strict_tpyes=1)".

Then we  require the "autoload.php" file.

Then we call the "Dotenv" object and the "$dotenv->load()" method.

Then a new "Database " object is is instantiated and the "$_ENV[]"  database credentials are passed in as arguments.

Then a new "RefreshTokenGateway" object is instantiated and the "$database" with the "$_ENV[]" secret key is passed in as arguments.

Lastly the "deleteExpiredToken()" method is called from the "RefreshTokenGateway" object to delete the expired "refresh_token" in the database.

Note: php version 8.2.0 is used for this code base.